backend/requirements.txt
pandas
numpy
scikit-learn
joblib
mlflow
SQLAlchemy
psycopg2-binary
fastapi
pydantic
uvicorn
xgboost
imbalanced-learn
Flask
flask-cors
statsmodels

pmdarima
scipy
matplotlib
python-dotenv
groq